refactor: remove obsolete remote agent query fallback
